Add functionality to handle copy and paste of article link, and pressing enter key to submit the form, and fix RTK bug in article service endpoint in Demo.jsx and article.js files

+++ src/components/Demo.jsx
@@ -26,15 +26,31 @@ const Demo = () => {
 
   const handleSubmit = async (e) => {
     e.preventDefault();
+    const existingArticle = allArticles.find(
+      (item) => item.url === article.url
+    );
+    if (existingArticle) return setArticle(existingArticle);
     const { data } = await getSummary({ articleUrl: article.url });
-
-    if(data?.summary) {
+    if (data?.summary) {
       const newArticle = { ...article, summary: data.summary };
       const updatedAllArticles = [newArticle, ...allArticles];
 
+      // update state and local storage
       setArticle(newArticle);
       setAllArticles(updatedAllArticles);
       localStorage.setItem("articles", JSON.stringify(updatedAllArticles));
+    }
+  };
+
+  const handleCopy = (copyUrl) => {
+    setCopied(copyUrl);
+    navigator.clipboard.writeText(copyUrl);
+    setTimeout(() => setCopied(""), 2000);
+  }
+
+  const handleKeyDown = (e) => {
+    if(e.keyCode === 13) {
+      handleSubmit(e);
     }
   }
 
@@ -53,15 +69,16 @@ const Demo = () => {
           />
 
           <input
+            type='url'
+            placeholder='Paste the article link'
             value={article.url}
             onChange={(e) => setArticle({ ...article, url: e.target.value })}
+            onKeyDown={handleKeyDown}
             required
-            className="url_input peer"
-            placeholder="Enter a URL"
-            type="url"
+            className='url_input peer' // When you need to style an element based on the state of a sibling element, mark the sibling with the peer class, and use peer-* modifiers to style the target element
           />
           <button
-            type="submit"
+            type='submit'
             className='submit_btn peer-focus:border-gray-700 peer-focus:text-gray-700 '
           >
             <p>↵</p>
@@ -91,6 +108,7 @@ const Demo = () => {
         </div>
       </div>
 
+      {/* Display Result */}
       <div className='my-10 max-w-full flex justify-center items-center'>
         {isFetching ? (
           <img src={loader} alt='loader' className='w-20 h-20 object-contain' />
@@ -108,8 +126,10 @@ const Demo = () => {
               <h2 className='font-satoshi font-bold text-gray-600 text-xl'>
                 Article <span className='blue_gradient'>Summary</span>
               </h2>
-              <div className="summary_box">
-                <p className="font-inter font-medium text-sm text-gray-700">{article.summary}</p>
+              <div className='summary_box'>
+                <p className='font-inter font-medium text-sm text-gray-700'>
+                  {article.summary}
+                </p>
               </div>
             </div>
           )
